+++ wemake_python_styleguide/violations/complexity.py
@@ -665,7 +665,7 @@ class TooManyConditionsViolation(ASTViolation):
     """
     Forbid conditions with too many logical operators.
 
-    We use :str:`wemake_python_styleguide.constants.MAX_CONDITIONS`
+    We use :str:`wemake_python_styleguide.options.defaults.MAX_CONDITIONS`
     as a default value.
 
     Reasoning:
@@ -678,11 +678,17 @@ class TooManyConditionsViolation(ASTViolation):
         creating new variables or creating nested ``if`` statements.
         Both of these actions will trigger other complexity checks.
 
+    Configuration:
+        This rule is configurable with ``--max-conditions``.
+        Default:
+        :str:`wemake_python_styleguide.options.defaults.MAX_CONDITIONS`
+
     We count ``and`` and ``or`` keywords as conditions.
 
     .. versionadded:: 0.1.0
     .. versionchanged:: 0.5.0
-
+    .. versionchanged:: 1.4.0
+        Added ``--max-conditions`` configuration options.
     """
 
     error_template = 'Found a condition with too much logic: {0}'